Applied patch from Joe Eckard to fix problem when there is more than one directory under the uploads directory; Jira #OFBIZ-318
git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@5283 deaff617-5be1-0310-a1b7-ec6878f92ba9

+++ applications/content/src/org/ofbiz/content/data/DataResourceWorker.java
@@ -996,9 +996,9 @@ public class DataResourceWorker {
         // descending comparator
         Comparator desc = new Comparator() {
             public int compare(Object o1, Object o2) {
-                if (((Integer) o1).intValue() > ((Integer) o2).intValue()) {
+                if (((Long) o1).longValue() > ((Long) o2).longValue()) {
                     return -1;
-                } else if (((Integer) o1).intValue() < ((Integer) o2).intValue()) {
+                } else if (((Long) o1).longValue() < ((Long) o2).longValue()) {
                     return 1;
                 }
                 return 0;
